change file organization
- Renamed borehole.py to boreholeUI.py
- Created borehole.py to hold class Borehole that will contain the
actual borehole data
- Added Contributors section to README
